Step 1: Create state/index.js core infrastructure
Move all state logic from state.js into state/index.js. Add mutate()
wrapper with monotonic version counter and getVersion(). Replace
state.js with a thin re-export shim so all existing imports work
unchanged.

- state/index.js: full state module + mutate/getVersion
- state/index.test.js: core infrastructure tests
- state.js: re-export shim
- Makefile: lint includes state/index.js, test glob includes js/**
