+++ Makefile
@@ -9,14 +9,14 @@ serve:
 # Excludes main.js (@ts-nocheck — CustomEvent.detail typing noise)
 lint:
 	npx tsc --noEmit --allowJs --checkJs --target ES2020 --moduleResolution bundler --module ES2020 \
-		js/types.js js/events.js js/state.js js/exploits.js js/combat.js js/loot.js \
+		js/types.js js/events.js js/state.js js/state/index.js js/exploits.js js/combat.js js/loot.js \
 		js/alert.js js/timers.js js/ice.js js/log.js js/log-renderer.js js/visual-renderer.js js/store.js js/console.js js/cheats.js \
 		js/node-types.js js/node-lifecycle.js \
 		js/node-actions.js js/global-actions.js js/action-context.js
 
 # Run unit + integration tests
 test:
-	node --test tests/*.test.js
+	node --test tests/*.test.js js/**/*.test.js
 
 # Full check: lint + test
 check: lint test
